pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "fraud-prediction-pipeline"
version = "0.1.0"
description = "End-to-end ML pipeline for fraud detection"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.0",
    "numpy>=1.24",
    "scikit-learn>=1.4",
    "xgboost>=2.0",
    "lightgbm>=4.0",
    "optuna>=3.5",
    "imbalanced-learn>=0.12",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "pydantic>=2.6",
    "pydantic-settings>=2.2",
    "mlflow>=2.12",
    "joblib>=1.3",
    "pyarrow>=15.0",
    "pyyaml>=6.0",
    "rich>=13.0",
    "typer>=0.12",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.9",
    "pre-commit>=3.7",
    "ipykernel>=6.0",
    "matplotlib>=3.8",
    "seaborn>=0.13",
    "shap>=0.45",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=src --cov-report=term-missing"

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
